+++ examples/typed/group/main.go
@@ -41,7 +41,7 @@ func main() {
 
 	// --- GET ---
 	fetchedGroup, err := rest.Groups.Get(&typed.GroupSearchParams{
-		Name: expr.S("go-client-test-group"),
+		Name: expr.Str("go-client-test-group"),
 	})
 	if err != nil {
 		panic(fmt.Errorf("failed to get group: %w", err))
@@ -61,7 +61,7 @@ func main() {
 
 	// --- DELETE ---
 	err = rest.Groups.Delete(&typed.GroupSearchParams{
-		Name: expr.S("go-client-test-group"),
+		Name: expr.Str("go-client-test-group"),
 	})
 	if err != nil {
 		panic(fmt.Errorf("failed to delete group: %w", err))